requirements.txt
appnope==0.1.2
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
black==21.9b0
bleach==4.1.0
cffi==1.14.6
click==8.0.3
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.1.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
ninja==1.10.2.2
notebook==6.4.4
numpy==1.21.2
packaging==21.0
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.10.8
Send2Trash==1.8.0
six==1.16.0
terminado==0.12.1
testpath==0.5.0
tomli==1.2.1
torch==1.9.1
torchvision==0.10.1
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wandb
omegaconf
scikit-image